+++ src/main.js
@@ -6,6 +6,10 @@ import Buefy from "buefy";
 import "buefy/dist/buefy.css";
 import "bulma-helpers/css/bulma-helpers.min.css";
 
+import Vue from 'vue'
+import axios from 'axios'
+import VueAxios from 'vue-axios'
+
 // Local files
 // Layouts
 import DefaultLayout from "~/layouts/Default.vue";
@@ -17,9 +21,12 @@ export default function(Vue, { router, head, isClient }) {
         rel: "stylesheet",
         href: "https://use.fontawesome.com/releases/v5.2.0/css/all.css"
     });
+
     Vue.use(Buefy, {
         defaultIconPack: "fas" // Font Awesome Solid
     });
+
+    Vue.use(VueAxios, axios);
 
     // Set default layout as a global component
     Vue.component("Layout", DefaultLayout);